Moving the exception-handlers to the domain folder
