--- src/main/java/com/ros/lms/adapters/inbound/exception_handlers/MemberControllerExceptionHandler.java
@@ -1,62 +1,0 @@
-package com.ros.lms.adapters.inbound.exception_handlers;
-
-import com.ros.lms.domain.exceptions.EmailAlreadyExistsException;
-import com.ros.lms.domain.exceptions.MemberAlreadyExistsException;
-import com.ros.lms.domain.exceptions.UsernameAlreadyExistsException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.ConstraintViolation;
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class MemberControllerExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<ProblemDetail> handleMemberAlreadyExistsException(MemberAlreadyExistsException ex, HttpServletRequest request){
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.CONFLICT, ex.getMessage());
-        problemDetail.setTitle("Existing Member Error");
-        problemDetail.setInstance(URI.create(request.getRequestURI()));
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ProblemDetail> handleUsernameAlreadyExistsException(UsernameAlreadyExistsException ex, HttpServletRequest request){
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.CONFLICT, ex.getMessage());
-        problemDetail.setTitle("Existing Username Error");
-        problemDetail.setInstance(URI.create(request.getRequestURI()));
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ProblemDetail> handleEmailAlreadyExistsException(EmailAlreadyExistsException ex, HttpServletRequest request){
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.CONFLICT, ex.getMessage());
-        problemDetail.setTitle("Existing Email Error");
-        problemDetail.setInstance(URI.create(request.getRequestURI()));
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(problemDetail);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<ProblemDetail> handleConstraintViolationException(ConstraintViolationException ex, HttpServletRequest request){
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setTitle("Validation Error");
-        problemDetail.setDetail("Invalid input data.");
-        problemDetail.setInstance(URI.create(request.getRequestURI()));
-
-        // Add field-specific validation details
-        Map<String, String> violations = new HashMap<>();
-        for (ConstraintViolation<?> violation : ex.getConstraintViolations()) {
-            violations.put(violation.getPropertyPath().toString(), violation.getMessage());
-        }
-        problemDetail.setProperty("violations", violations);
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
-    }
-}
