+++ src/main/java/com/ros/lms/domain/exception_handlers/StaffExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.ros.lms.adapters.inbound.exception_handlers;
+package com.ros.lms.domain.exception_handlers;
 
-import com.ros.lms.domain.exceptions.StorageFileNotFoundException;
+import com.ros.lms.domain.exceptions.StaffNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
@@ -11,12 +11,12 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import java.net.URI;
 
 @ControllerAdvice
-public class StorageControllerExceptionHandler {
+public class StaffExceptionHandler {
 
     @ExceptionHandler
-    public ResponseEntity<ProblemDetail> handleStorageFileNotFoundExceptionException(StorageFileNotFoundException ex, HttpServletRequest request){
+    public ResponseEntity<ProblemDetail> handleStaffNotFoundException(StaffNotFoundException ex, HttpServletRequest request) {
         ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, ex.getMessage());
-        problemDetail.setTitle("File not Found Error");
+        problemDetail.setTitle("Staff not Found");
         problemDetail.setInstance(URI.create(request.getRequestURI()));
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(problemDetail);
     }
