+++ src/main/java/com/ros/lms/domain/exception_handlers/StorageExceptionHandler.java
@@ -1,0 +1,32 @@
+package com.ros.lms.domain.exception_handlers;
+
+import com.ros.lms.domain.exceptions.StorageException;
+import com.ros.lms.domain.exceptions.StorageFileNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ProblemDetail;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.net.URI;
+
+@ControllerAdvice
+public class StorageExceptionHandler {
+
+    @ExceptionHandler
+    public ResponseEntity<ProblemDetail> handleStorageFileNotFoundExceptionException(StorageFileNotFoundException ex, HttpServletRequest request){
+        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, ex.getMessage());
+        problemDetail.setTitle("File not Found Error");
+        problemDetail.setInstance(URI.create(request.getRequestURI()));
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(problemDetail);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<ProblemDetail> handleStorageException(StorageException ex, HttpServletRequest request) {
+        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, ex.getMessage());
+        problemDetail.setTitle("Storage Error");
+        problemDetail.setInstance(URI.create(request.getRequestURI()));
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(problemDetail);
+    }
+}
