+++ src/main/resources/application-prod.properties
@@ -4,7 +4,10 @@ logging.level.org.springframework.web=DEBUG
 logging.level.org.springframework.security=DEBUG
 
 cloud.aws.region.static=us-east-1
-spring.config.import=aws-parameterstore:/config/ros-lms-db,aws-parameterstore:/config/ros-lms-keys,aws-parameterstore:/config/ros-lms-app
+spring.config.import=\
+  aws-parameterstore:/config/ros-lms-db,\
+  aws-parameterstore:/config/ros-lms-keys,\
+  aws-parameterstore:/config/ros-lms-app
 
 
 #This property is needed to read the encrypted passwords in the database
